server-side/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
appwrite==11.1.0